Added files via upload
Okay, to be honest, I have no idea what I'm doing with this. I just downloaded the level and changed its name after making additions. So far I've just had time to expand the rooms, add upstairs access to their roofs and throw in corridor access to the towers. I'll add in stairs and walls for the maze as soon as I can.
Also, I figured out what was up with the floor. It was overlapping with the other floor. I've lowered the floor a little bit and fixed this issue.